requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.32.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.1
langgraph>=0.0.1
langchain>=0.1.0
langchain-openai>=0.3.9
qdrant-client>=1.7.0
chromadb>=0.4.22
langchain-qdrant>=0.0.3
firecrawl-py>=1.15.0
-e .